:truck: :: move snap dto
