+++ src/main/kotlin/com/photo/server/starsnap/domain/snap/dto/SnapDto.kt
@@ -1,4 +1,4 @@
-package com.photo.server.starsnap.domain.snap.controller.dto
+package com.photo.server.starsnap.domain.snap.dto
 
 import com.photo.server.starsnap.global.dto.SnapDto
 import com.photo.server.starsnap.global.dto.SnapUserDto
