+++ src/main/kotlin/com/photo/server/starsnap/domain/snap/service/SnapAwsS3Service.kt
@@ -42,7 +42,7 @@ class SnapAwsS3Service(
         }
     }
 
-    fun fixImage(image: MultipartFile, imageKey: String) {
+    fun updateImage(image: MultipartFile, imageKey: String) {
         if (doesObjectExist(imageKey)) {
             val imageSize = image.size
